unsigned int --> int

+++ src/Classic/Structure/PeakFinder.cpp
@@ -308,7 +308,7 @@ void PeakFinder::createHistogram_m() {
     for(container_t::iterator it = radius_m.begin(); it != radius_m.end(); ++it) {
         int bin = std::abs(*it - globMin_m ) * invBinWidth;
         
-        if ( bin >= locHist.size() )
+        if ( bin >= (int)locHist.size() )
              bin = locHist.size() - 1;
         
         ++locHist[bin];
